refactor: add utils directories
