+++ src/components/Cells/DayCells/DayCells.tsx
@@ -6,10 +6,10 @@ import { ColorContext } from '@/context/colorContext';
 import { DateContext } from '@/context/dateContext';
 import { WeekContext } from '@/context/weekContext';
 import { Day } from '@/types/Day';
+import { isRangeDate } from '@/utils/dates/isRangeDate';
+import { isSameDates } from '@/utils/dates/isSameDates';
 import { getDaysByCalendarType } from '@/utils/dayCells/getDaysByCalendarType';
 import { hasTodos } from '@/utils/dayCells/hasTodos';
-import { isRangeDate } from '@/utils/isRangeDate';
-import { isSameDates } from '@/utils/isSameDates';
 
 export interface DayCellsProps {
   days: Day[];
