+++ src/hocs/withRangeCalendarLogic.tsx
@@ -6,8 +6,8 @@ import { DateContext } from '@/context/dateContext';
 import { useCalendarContext } from '@/hooks/useCalendarContext';
 import { useDays } from '@/hooks/useDays';
 import { ConfigurableElementProps } from '@/types/ConfigurableElementProps';
+import { isSameDates } from '@/utils/dates/isSameDates';
 import { isCheckedRangeCell } from '@/utils/dayCells/isCheckedRangeCell';
-import { isSameDates } from '@/utils/isSameDates';
 
 export const withRangeCalendarLogic = <T extends ConfigurableElementProps>(
   WrappedComponent: ComponentType<T>,
