+++ src/utils/dayCells/hasTodos.ts
@@ -1,4 +1,4 @@
-import { getTodosByDate } from '@/utils/getTodosByDate';
+import { getTodosByDate } from '@/utils/todos/getTodosByDate';
 
 export const hasTodos = (year: number, month: number, selectedDay: number) => {
   const todosList = getTodosByDate(new Date(year, month, selectedDay));
