+++ src/utils/helpers/deleteEveryNthElement.ts
