+++ src/utils/helpers/shiftArrayToLeft.ts
@@ -1,4 +1,4 @@
-import { shiftArray } from '@/utils/shiftArray';
+import { shiftArray } from '@/utils/helpers/shiftArray';
 
 export const shiftArrayToLeft = <T>(arr: T[], shiftCount: number): T[] =>
   shiftArray(
